+++ addons/yat/src/commands/builtin/View.cs
@@ -1,5 +1,6 @@
 using System;
 using Godot;
+using YAT.Overlay.Components.Terminal;
 
 namespace YAT.Commands
 {