+++ addons/yat/src/helpers/LogHelper.cs
@@ -1,4 +1,5 @@
 using Godot;
+using YAT.Overlay.Components.Terminal;
 
 namespace YAT.Helpers
 {